requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2 # для работы django-rest-knox
defusedxml==0.7.1
Django==4.2
django-rest-knox==4.2.0 # jwt
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
